+++ Model/essLinacInc/FaradayCup.h
@@ -75,7 +75,7 @@ class FaradayCup : public attachSystem::ContainedComp,
   std::vector<int>    shieldMat;   ///< shielding material
 
   void layerProcess(Simulation& System, const std::string& cellName,
-		    const size_t& lpS, const size_t& lsS, const size_t&, const int&);
+		    const long int& lpS, const long int& lsS, const size_t&, const int&);
 
   void populate(const FuncDataBase&);
   void createUnitVector(const attachSystem::FixedComp&,
